requirements.txt
fastapi
uvicorn
python-multipart
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-qdrant==0.2.1
langgraph
openai>=1.58.1
qdrant-client>=1.12.0
sqlite-utils
pydantic
tiktoken
python-dotenv
toml
pandas
numpy
requests
SQLAlchemy
streamlit